+++ packages/adminui-vue/components/navbar/UNavbar.vue
@@ -109,7 +109,7 @@ export default {
       tabs: [],
       activeTabId: null,
       contextMenuTabId: null,
-      isCollapsed: window.localStorage.getItem('portal:sidebar:isCollapsed') === 'true',
+      isCollapsed: this.$uiSettings.get('sidebar', 'isCollapsed') === true,
       visibleNavbar: true,
       originalExtNavbarHeight: null
     }
@@ -117,12 +117,13 @@ export default {
 
   watch: {
     visibleNavbar (value) {
+      const centralPanel = this.$UB.core.UBApp.viewport.centralPanel
       if (value) {
-        this.$UB.core.UBApp.viewport.centralPanel.tabBar.setHeight(this.originalExtNavbarHeight)
-        this.$UB.core.UBApp.viewport.centralPanel.setMargin(`-${this.originalExtNavbarHeight} 0 0 0`)
+        centralPanel.tabBar.setHeight(this.originalExtNavbarHeight)
+        centralPanel.setMargin(`-${this.originalExtNavbarHeight} 0 0 0`)
       } else {
-        this.$UB.core.UBApp.viewport.centralPanel.tabBar.setHeight(0)
-        this.$UB.core.UBApp.viewport.centralPanel.setMargin(`-${this.originalExtNavbarHeight} 0 0 0`)
+        centralPanel.tabBar.setHeight(0)
+        centralPanel.setMargin(`-${this.originalExtNavbarHeight} 0 0 0`)
       }
     }
   },
@@ -133,12 +134,13 @@ export default {
   },
 
   mounted () {
-    this.tabs = window.$App.viewport.centralPanel.items.items
+    const centralPanel = this.$UB.core.UBApp.viewport.centralPanel
+    this.tabs = centralPanel.items.items
       .map(({ id, title, titleTooltip }) => ({ id, title, titleTooltip }))
 
     this.originalExtNavbarHeight = this.$el.offsetHeight
-    this.$UB.core.UBApp.viewport.centralPanel.tabBar.setHeight(this.originalExtNavbarHeight)
-    this.$UB.core.UBApp.viewport.centralPanel.setMargin(`-${this.originalExtNavbarHeight} 0 0 0`)
+    centralPanel.tabBar.setHeight(this.originalExtNavbarHeight)
+    centralPanel.setMargin(`-${this.originalExtNavbarHeight} 0 0 0`)
   },
 
   methods: {