+++ packages/uba/uba_user.js
@@ -43,7 +43,7 @@ function checkDuplicateUser (ctxt) {
       .select()
 
     if (!store.eof) {
-      throw new UB.UBAbort('<<<Duplicate user name (may be in different case)>>>')
+      throw new UB.UBAbort('<<<uba_user_errors.duplicateUserName>>>')
     }
     params.name = newName.toLowerCase().trim() // convert a username to lower case
   }
@@ -527,11 +527,11 @@ function denyBuildInUserDeletion (ctx) {
   const userName = ctx.dataStore.get('name')
 
   if (isBuiltInUser(userName)) {
-    throw new UB.UBAbort('<<<Removing of built-in user is prohibited>>>')
+    throw new UB.UBAbort(UB.i18n('<<<uba_user.deleteBuiltInUserProhibited>>>'))
   }
 
   if (isDisabledUser(userName)) {
-    throw new UB.UBAbort('<<<Removing of service user is prohibited>>>')
+    throw new UB.UBAbort(UB.i18n('<<<uba_user.deleteServiceUserProhibited>>>'))
   }
 }
 
@@ -555,11 +555,11 @@ function denyBuildInUserRename (ctx) {
   }
 
   if (isBuiltInUser(userName)) {
-    throw new UB.UBAbort('<<<Renaming of built-in user is prohibited>>>')
+    throw new UB.UBAbort(UB.i18n('<<<uba_user.renameBuiltInUserProhibited>>>'))
   }
 
   if (isDisabledUser(userName)) {
-    throw new UB.UBAbort('<<<Renaming of service user is prohibited>>>')
+    throw new UB.UBAbort(UB.i18n('<<<uba_user.renameServiceUserProhibited>>>'))
   }
 }